backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
distro==1.8.0
Django==3.2.18
django-cors-headers==3.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.0
paramiko==3.0.0
Pillow==9.4.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.14
websocket-client==0.59.0
